+++ resources/views/livewire/tables/units-table.blade.php
@@ -75,26 +75,33 @@
                     @unless($good)
                         <x-tables.td>
                             <x-form.input.toggle wire:change="toggleActive({{ $unit->id }})"
-                                                 name="{{ $unit->id }}" :isActive="$unit->is_active" wire:key="toggle-{{ $unit->id }}"/>
+                                                 name="{{ $unit->id }}" :isActive="$unit->is_active" wire:key="toggle-{{ $unit->id }}"
+                                                 :disabled="! auth()->user()?->can('units.edit')"/>
                         </x-tables.td>
                     @endunless
                     <x-tables.td class="flex space-x-2">
                         <x-tables.action-show :route="route('units.show', $unit->id)"/>
                         @unless($readonly)
-                            <x-tables.action-edit :route="route('units.edit', $unit->id)"/>
+                            @can('units.edit')
+                                <x-tables.action-edit :route="route('units.edit', $unit->id)"/>
+                            @endcan
                             @if($good)
-                                <x-tables.action-delete
-                                    wire:click="deleteUnit({{ $unit->id }})"
-                                    :confirm="__('goods.confirm_remove_unit_assignment')"
-                                    :label="__('goods.remove_unit_assignment')"
-                                >
-                                    <x-heroicon-o-link-slash class="w-6 h-6 hover:text-red-500"/>
-                                </x-tables.action-delete>
+                                @can('goods.edit')
+                                    <x-tables.action-delete
+                                        wire:click="deleteUnit({{ $unit->id }})"
+                                        :confirm="__('goods.confirm_remove_unit_assignment')"
+                                        :label="__('goods.remove_unit_assignment')"
+                                    >
+                                        <x-heroicon-o-link-slash class="w-6 h-6 hover:text-red-500"/>
+                                    </x-tables.action-delete>
+                                @endcan
                             @else
-                                <x-tables.action-delete
-                                    wire:click="deleteUnit({{ $unit->id }})"
-                                    :confirm="__('units.confirm_delete_unit')"
-                                />
+                                @can('units.delete')
+                                    <x-tables.action-delete
+                                        wire:click="deleteUnit({{ $unit->id }})"
+                                        :confirm="__('units.confirm_delete_unit')"
+                                    />
+                                @endcan
                             @endif
                         @endunless
                     </x-tables.td>
