+++ resources/views/livewire/tables/vehicles-table.blade.php
@@ -95,15 +95,20 @@
                     <x-tables.td><x-ui.expiry-date-badge :date="$vehicle->tachograph_inspection_expiry_date"/></x-tables.td>
                     <x-tables.td>
                         <x-form.input.toggle wire:change="toggleActive({{ $vehicle->id }})"
-                                             name="{{ $vehicle->id }}" :isActive="$vehicle->is_active" wire:key="toggle-{{ $vehicle->id }}"/>
+                                             name="{{ $vehicle->id }}" :isActive="$vehicle->is_active" wire:key="toggle-{{ $vehicle->id }}"
+                                             :disabled="! auth()->user()?->can('vehicles.edit')"/>
                     </x-tables.td>
                     <x-tables.td class="flex space-x-2">
                         <x-tables.action-show :route="route('vehicles.show', $vehicle->id)"/>
-                        <x-tables.action-edit :route="route('vehicles.edit', $vehicle->id)"/>
-                        <x-tables.action-delete
-                            wire:click="deleteVehicle({{ $vehicle->id }})"
-                            :confirm="__('vehicles.confirm_delete_vehicle')"
-                        />
+                        @can('vehicles.edit')
+                            <x-tables.action-edit :route="route('vehicles.edit', $vehicle->id)"/>
+                        @endcan
+                        @can('vehicles.delete')
+                            <x-tables.action-delete
+                                wire:click="deleteVehicle({{ $vehicle->id }})"
+                                :confirm="__('vehicles.confirm_delete_vehicle')"
+                            />
+                        @endcan
                     </x-tables.td>
                 </tr>
             @endforeach
